структурировать тесты сервисного слоя по пакетам
