+++ src/test/java/ru/practicum/shareit/repository/CommentRepositoryTest.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.repository;
+
+public class CommentRepositoryTest {
+}
