+++ src/test/java/ru/practicum/shareit/repository/ItemRepositoryTest.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.repository;
+
+public class ItemRepositoryTest {
+}
