+++ src/test/java/ru/practicum/shareit/repository/RequestRepositoryTest.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.repository;
+
+public class RequestRepositoryTest {
+}
