+++ src/test/java/ru/practicum/shareit/service/integration/ItemRequestIntegrationTest.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.service.integration;
+
+public class ItemRequestIntegrationTest {
+}
